refactor: simplfy, remove conditional provisioning
I've opted to use individual branches to convey specific points in this
story instead of conditionally provisioning resources.

This way each story can be a little simpler and more tightly coupled
with the article so that things don't change at some future date unless
I explicitly intend for them to.